requirements.txt
discord.py==2.3.2
gspread==5.12.0
oauth2client==4.1.3
pytz==2023.3